requirements.txt
scipy==0.15.0
numpy==1.9.1
scikit-learn==0.15.2
librosa==0.3.1
mixpanel-py==3.2.1
